+++ CHANGELOG.md
@@ -16,6 +16,9 @@ You can also check the
     parameter
   - It's now possible to add custom WMS layers coming from wms.geo.admin.ch to
     map charts
+  - Statistics page was enhanced with `Charts` and `Cubes` sections, including
+    some general improvements
+  - Application now collect data on chart previews, not only views
 - Fixes
   - Limits that are connected to values that are filtered out do not render in
     the chart anymore